requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# MLOps
mlflow==2.9.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0